Project_Files/requirements.txt
requests
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
joblib
python-pptx
streamlit
